% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{args_as_list}
\alias{args_as_list}
\title{Coerce \code{...} to List}
\usage{
args_as_list(...)
}
\arguments{
\item{...}{Arguments.}
}
\value{
A list of arguments.
}
\description{
Coerce \code{...} to List
}
