% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-consensus.R
\name{as_consensus_cluster}
\alias{as_consensus_cluster}
\title{Assign Consensus Cluster Labels}
\usage{
as_consensus_cluster(mat, wildcard_cutoff = 0.5, ref_col = NA, ...)
}
\arguments{
\item{mat}{A matrix or data.frame.}

\item{wildcard_cutoff}{A number between 0 and 1. Sets the minimum
proportion of identical labels needed for a row to be assigned a
final label. For example, if \code{wildcard_cutoff = 0.5}, then a row
must be assigned the label "A" in half or more of all columns to be
assigned "A" as the final label. If the most common label for a row
appears less frequently than the cutoff, the row is given an NA
label. Use \code{wildcard_cutoff = 0} to disable this feature.
When the cutoff is met, the statistical mode is used. All ties
are broken randomly.}

\item{ref_col}{An integer. The column to use as a reference.
If NA, the function will choose a reference automatically.}

\item{...}{Arguments to \code{\link{match_labels}}.}
}
\value{
A vector of new labels.
}
\description{
First, this function builds a "consensus matrix" to give different
 clustering results a common vocabulary so that they can be compared
 with one another. Second, this function assigns each row a final
 label based on which label is most common. Rows that do not have
 consistent labels across the "consensus matrix" are considered
 to be "wild cards", and given an NA label.
}
