% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-consensus.R
\name{as_consensus_matrix}
\alias{as_consensus_matrix}
\title{Build Consensus Matrix}
\usage{
as_consensus_matrix(mat, ref_col = NA, ...)
}
\arguments{
\item{mat}{A matrix or data.frame.}

\item{ref_col}{An integer. The column to use as a reference.
If NA, the function will choose a reference automatically.}

\item{...}{Arguments to \code{\link{match_labels}}.}
}
\value{
A data.frame of matched labels.
}
\description{
This function runs \code{\link{match_labels}} on every column
 of a matrix. It can be used to give different clustering results a
 common vocabulary so that they can be compared with one another.
 In other words, this function establishes a "consensus matrix"
 that can be used for consensus clustering.
}
