% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{as_factor_frame}
\alias{as_factor_frame}
\title{Coerce Data Frame Into "Factor Frame"}
\usage{
as_factor_frame(data)
}
\arguments{
\item{data}{Any matrix or data.frame.}
}
\value{
A data.frame where each column is a factor.
}
\description{
This backend function coerces any matrix or data.frame into a
 new data.frame where each column is a factor.
}
