% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-circularize.R
\name{circularize}
\alias{circularize}
\title{Circularize a Difference Measure}
\usage{
circularize(FUN, max_shift, return_min = TRUE)
}
\arguments{
\item{FUN}{The function to circularize.}

\item{max_shift}{An integer. The number of alignments to try.
When \code{max_shift = 1}, only the default alignment is used.}

\item{return_min}{A logical. Toggles whether the function
should return all difference measures, or the smallest.}
}
\value{
A function. An instance of \code{FUN} that now
 considers multiple alignments.
}
\description{
The difference measures included in this package compare two
 factor-class vectors of the same size. The measures do not consider
 any different alignments of the vectors. With \code{circularize}, a
 difference function is made to consider multiple alignments.
}
