% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-diffs.R
\name{d_adjRand}
\alias{d_adjRand}
\title{Calculate Adjusted Rand}
\usage{
d_adjRand(a, b)
}
\arguments{
\item{a, b}{Factors.}
}
\value{
The difference between \code{a} and \code{b}.
}
\description{
A factor metric that is categorical and relative.
}
\details{
See Section 2.2.1 of Caruana et al. paper "Meta Clustering".
 This metric sums I_ij for all pairs (i,j) where I_ij=1 if points i and j
 have the same label in one input but different labels in the other input.
 The total sum is then scaled by the total number of pairs, and so the
 metric will always range from 0 to 1.
}
