% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-diffs-jaccard.R
\name{d_jaccard}
\alias{d_jaccard}
\title{Calculate Jaccard Distance}
\usage{
d_jaccard(a, b)
}
\arguments{
\item{a, b}{A vector, typically a factor.}
}
\value{
The difference between \code{a} and \code{b}.
}
\description{
A factor metric that is categorical and absolute.
}
\details{
This function measures the Jaccard distance between the elements of
 \code{a} and the elements of \code{b}. It normalizes the size of
 the set union by the size of the individual sets.

NAs have no category assignment. As such, an NA never equals an NA.
 Two vectors of NAs are maximally different because they never equal
 one another. A vector of NAs and a vector of a single category is
 also maximally different because they never equal one another.
 NAs are ignored in the denominator of Jaccard. So, the input
 \code{a=c("A", "A", NA, NA)} is the same as
 \code{b=c("A", "A", NA, NA)}.
}
