% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{do_call}
\alias{do_call}
\title{A Simple(r) \code{do.call} Function}
\usage{
do_call(FUN, ..., .prior_arg_list = list())
}
\arguments{
\item{FUN}{A function to call.}

\item{...}{Arguments.}

\item{.prior_arg_list}{More arguments. Used to pass an argument
list collected from \code{\link{args_as_list}}. Unlike \code{...},
this can be used to pass empty arguments, by setting
\code{.prior_arg_list = list()}.}
}
\value{
The result of \code{FUN}.
}
\description{
A Simple(r) \code{do.call} Function
}
