% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-factordist.R
\name{factordist}
\alias{factordist}
\title{Calculate Distance Between Factors}
\usage{
factordist(data, metric = c("jaccard", "adjRand"), as_dist = TRUE)
}
\arguments{
\item{data}{A data.frame of factors. Distances and differences
are calculated between columns of the input. If data does not contain
factors, factors are introduced by \code{\link{as_factor_frame}}.}

\item{metric}{A string. The name of the difference metric.}

\item{as_dist}{A boolean. Toggles whether to return the
difference matrix as a Base R \code{dist} object.}
}
\value{
A difference matrix.
}
\description{
Calculate Distance Between Factors
}
\details{
A metric can be categorical, ranked, absolute, or relative.
 Categorical metrics treat input as unranked characters.
 Ranked metrics treat input as ranked factors.
 Absolute metrics compare factor labels in one input directly
  with factor labels in the other input. For example, a=c(1,1,2,2) is
  different from b=c(2,2,1,1) because 1 is not the same as 2.
 Relative metrics compare groups of factor labels in one input
  with groups of factor labels in the other input. For example,
  a=c(1,1,2,2) is the same as b=c(2,2,1,1) because the first and
  second index, as well as the third and fourth index, always
  have the same label.
}
\examples{
library(factordist)
x <- simulate_factor_data(30, 6)
dis <- factordist(x, metric = "adjRand")

}
