% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-diffs-adjRand.R
\name{ij_same_any}
\alias{ij_same_any}
\title{Check Element Equivalency}
\usage{
ij_same_any(a)
}
\arguments{
\item{a}{A vector, typically a factor.}
}
\value{
A square matrix where \code{[i,j]=1} when
 \code{a[i] == a[j]}, or 0 otherwise.
}
\description{
Backend function, used by \code{d_adjRand}.
}
