% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-consensus.R
\name{match_labels}
\alias{match_labels}
\title{Match Labels}
\usage{
match_labels(a, b, strict = FALSE, verbose = FALSE)
}
\arguments{
\item{a, b}{Factors.}

\item{strict}{A boolean. Toggles whether the labels in \code{b} should
have a 1-to-1 mapping to the labels in \code{a}. If \code{FALSE}, the
multiple labels in \code{b} can map to one label in \code{a}.}

\item{verbose}{A boolean. Toggles whether to print intermediate results.}
}
\value{
A vector of new labels for \code{b}.
}
\description{
This function finds new labels for \code{b} that best match the labels
 of \code{a}. It can be used to give different clustering results a common
 vocabulary so that they can be compared with one another.
}
