% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{mode}
\alias{mode}
\title{Get Mode}
\usage{
mode(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The mode of the vector. If there are multiple modes,
 returns a mode randomly.
}
\description{
The mode will ignore NAs. If all elements are NA, the mode is NA.
}
