% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-papply.R
\name{papply}
\alias{papply}
\title{Apply Functions Over Pairs of Vectors}
\usage{
papply(X, MARGIN = 2, FUN, lower_only = FALSE, ...)
}
\arguments{
\item{X}{A matrix.}

\item{MARGIN}{Choose from 1 or 2. If 1, apply function to
pairs of rows. If 2, apply function to pairs of columns.}

\item{FUN}{The function to apply to row or column pairs.
The function should take two vectors as input.}

\item{lower_only}{A boolean. Toggles whether to apply
function to lower triangle pairs only.}

\item{...}{Arguments passed to \code{FUN}.}
}
\value{
A square matrix.
}
\description{
This function applies another function over pairs of rows
 or columns. It treats each row (or column) as a vector,
 then applies a function to each row pair (or column).
 For i,j in 1...N rows (or columns), it returns an N by N matrix
 containing the result of FUN(Xi, Xj).
}
\examples{
library(factordist)
x <- simulate_factor_data(30, 6)
mat <- papply(x, 2, d_adjRand)

}
