% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-factordist.R
\name{ready_data}
\alias{ready_data}
\title{Coerce Input As Factor}
\usage{
ready_data(data)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix}.}
}
\value{
A \code{data.frame} where each column is a \code{factor}.
}
\description{
This backend function coerces the input into a \code{data.frame}
 where each column is a \code{factor}.
}
