% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{resample}
\alias{resample}
\title{A Simple(r) \code{sample} Function}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Arguments to \code{sample}.}
}
\value{
A vector.
}
\description{
See \code{?sample}.
}
