% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-diffs-jaccard.R
\name{s_union}
\alias{s_union}
\title{Calculate Union Similarity}
\usage{
s_union(a, b)
}
\arguments{
\item{a, b}{A vector, typically a factor.}
}
\value{
The similarity between \code{a} and \code{b}.
}
\description{
A factor metric that is categorical and absolute.
}
\details{
This function counts the number of times that an element of \code{a}
 equals an element of \code{b}. In other words, \code{sum(a==b)}.

NAs have no category assignment. As such, an NA never equals an NA.
 Two vectors of NAs are maximally different because they never equal
 one another. A vector of NAs and a vector of a single category is
 also maximally different because they never equal one another.
 NAs are ignored in the denominator of Jaccard. So, the input
 \code{a=c("A", "A", NA, NA)} is the same as
 \code{b=c("A", "A", NA, NA)}.
}
