% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-simulate.R
\name{simulate_factor_data}
\alias{simulate_factor_data}
\title{Simulate Factor Data}
\usage{
simulate_factor_data(nrows, ncols, nletters = 26)
}
\arguments{
\item{nrows}{An integer. The number of rows (as samples).}

\item{ncols}{An integer. The number of columns (as variables).}

\item{nletters}{An integer. The number of letters.}
}
\value{
A data.frame where each column is a factor.
}
\description{
Simulate Factor Data
}
