% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-sym.R
\name{sym}
\alias{sym}
\title{Symmetrize Matrix}
\usage{
sym(mat)
}
\arguments{
\item{mat}{A square matrix.}
}
\value{
A square matrix.
}
\description{
Symmetrize a square matrix by copying the lower
 triangle to the upper triangle.
}
