% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-sym.R
\name{sym_min}
\alias{sym_min}
\title{Symmetrize Matrix}
\usage{
sym_min(mat)
}
\arguments{
\item{mat}{A square matrix.}
}
\value{
A square matrix.
}
\description{
Symmetrize a square matrix, but use the smallest
 of the lower and upper triangle values.
}
