% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{which_max}
\alias{which_max}
\title{Get Index of Maximum Value}
\usage{
which_max(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The index of the maximum of the vector. If there are
 multiple maximums, returns an index randomly.
}
\description{
Get Index of Maximum Value
}
