% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{which_min}
\alias{which_min}
\title{Get Index of Minimum Value}
\usage{
which_min(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The index of the minimum of the vector. If there are
 multiple minimum, returns an index randomly.
}
\description{
Get Index of Minimum Value
}
