% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-sim.R
\name{simulate_factor_data}
\alias{simulate_factor_data}
\title{Simulate Factor Data}
\usage{
simulate_factor_data(nrows, ncols)
}
\arguments{
\item{nrows}{The number of rows (as samples).}

\item{ncols}{The number of columns (as variables).}
}
\value{
A data.frame of factors.
}
\description{
Simulate Factor Data
}
